screen is the most common media type.
min width is to be read as 'starting from', while max width is to be read as up to.

it is better to use min width because of mobile first design, where the developer makes the mobile layout first before scaling it up to full browser size, as that is easier than scaling down.